+++ Pizzazz/pizzazz.h
@@ -228,25 +228,25 @@ namespace pizzazz {
        If the cursor is at the top of the window, this function will
        have no effect.
     */
-    void move_cursor_up(unsigned = 1);
+    void move_cursor_up(size_t = 1);
     
     /* Moves the terminal cursor down a chosen number of lines.
        If the cursor is at the bottom of the window, this function will
        have no effect.
     */
-    void move_cursor_down(unsigned = 1);
+    void move_cursor_down(size_t = 1);
     
     /* Moves the terminal cursor to the right a chosen number of columns.
        If the cursor is at the right edge of the window, this function will
        have no effect.
     */
-    void move_cursor_right(unsigned = 1);
+    void move_cursor_right(size_t = 1);
     
     /* Moves the terminal cursor to the left a chosen number of columns.
        If the cursor is at the left edge of the window, this function will
        have no effect.
     */
-    void move_cursor_left(unsigned = 1);
+    void move_cursor_left(size_t = 1);
     
     /* Detects the terminal window's current size in columns and rows. */
     Coord get_window_size();
